populate spinners

+++ app/src/main/java/com/example/lms/model/Course.java
@@ -8,47 +8,47 @@ import androidx.room.PrimaryKey;
 public class Course {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "course_id")
-    private int courseID;
+    private long courseID;
 
     @ColumnInfo(name = "grade_id")
-    private int gradeID;
+    private long gradeID;
 
     @ColumnInfo(name = "instructor_id")
-    private int instructorID;
+    private long instructorID;
 
     @ColumnInfo
     private String name;
     @ColumnInfo
     private String description;
 
-    public Course(int gradeID, int instructorID, String name, String description) {
+    public Course(long gradeID, long instructorID, String name, String description) {
         this.gradeID = gradeID;
         this.instructorID = instructorID;
         this.name = name;
         this.description = description;
     }
 
-    public int getCourseID() {
+    public long getCourseID() {
         return courseID;
     }
 
-    public void setCourseID(int courseID) {
+    public void setCourseID(long courseID) {
         this.courseID = courseID;
     }
 
-    public int getGradeID() {
+    public long getGradeID() {
         return gradeID;
     }
 
-    public void setGradeID(int gradeID) {
+    public void setGradeID(long gradeID) {
         this.gradeID = gradeID;
     }
 
-    public int getInstructorID() {
+    public long getInstructorID() {
         return instructorID;
     }
 
-    public void setInstructorID(int instructorID) {
+    public void setInstructorID(long instructorID) {
         this.instructorID = instructorID;
     }
